requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.19
botocore==1.31.19
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.12.2
frozenlist==1.4.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
networkx==3.1
ngrok==0.8.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
openai-whisper==20230314
protobuf==4.21.12
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
